Add TokenStack to replace Stack<Token>
Building is non-parallelizable, and Stack is thread safe, I've been seeing a lot of performance penalties with big expressions in highly threaded scenarios, after a lot of profiling the Vector implementation of Stack seems to be the culprit.

Why not use LinkedList, ArrayDeque, etc? they are overly general somewhat heavy clasess, I find that this is preferrable (in case it's not we can always rollback).