Added Expression#copy() and tests
This is actually following the "not-thread-safe" argument. The easiest
way to run this in a threaded scenario is simply to copy the Expression.

Since I did some changes in the way variables are stored I need to make
a copy of the VariableToken to avoid crossed reference assignment (see
testCopy2)